be stricter about the structure of a Facebook post
Other users' comments can also include <h3> and <table> tags. Currently
the best way to distinguish the poster's text is to assert that the divs
we're looking at are the first ones in the post. All of this might break
if Facebook changes its layout, but hopefully we can be strict at first
and loosen up later if need be.