Add eldest_kid checking in proof engine
Note about eldest_kid in Eldest proof:
We plan to make it not optional, but equal to kid being posted. So if an
Eldest proof is being posted with a key that becomes eldest, its
eldest_kid should point at said key. But Keybase client does not do this
right now, so this is postponed until client is changed and users get
the new client.